Move py3 only syntax to special files.
