+++ kivy/base.py
@@ -29,11 +29,16 @@ from kivy.clock import Clock
 from kivy.event import EventDispatcher
 from kivy.lang import Builder
 from kivy.context import register_context
-from kivy.compat import async_coroutine
 try:
     import asyncio
 except ImportError:
     asyncio = None
+
+if asyncio:
+    from kivy.base_async import AsyncEventLoopBase, async_runTouchApp
+else:
+    from kivy.compat import PY3CompatCls as AsyncEventLoopBase
+    async_runTouchApp = None
 
 # private vars
 EventLoop = None
@@ -98,7 +103,7 @@ class ExceptionManagerBase:
 ExceptionManager = register_context('ExceptionManager', ExceptionManagerBase)
 
 
-class EventLoopBase(EventDispatcher):
+class EventLoopBase(AsyncEventLoopBase, EventDispatcher):
     '''Main event loop. This loop handles the updating of input and
     dispatching events.
     '''
@@ -346,22 +351,6 @@ class EventLoopBase(EventDispatcher):
                 else:
                     pass
 
-    @async_coroutine
-    def async_mainloop(self):
-        while not self.quit and self.status == 'started':
-            try:
-                yield from self.async_idle()
-                if self.window:
-                    self.window.mainloop()
-            except BaseException as inst:
-                # use exception manager first
-                r = ExceptionManager.handle_exception(inst)
-                if r == ExceptionManager.RAISE:
-                    stopTouchApp()
-                    raise
-                else:
-                    pass
-
     def idle(self):
         '''This function is called after every frame. By default:
 
@@ -400,51 +389,11 @@ class EventLoopBase(EventDispatcher):
 
         return self.quit
 
-    @async_coroutine
-    def async_idle(self):
-        '''Identical to :meth:`idle`, but instead used when running
-        within an async event loop.
-        '''
-
-        # update dt
-        yield from Clock.async_tick()
-
-        # read and dispatch input from providers
-        self.dispatch_input()
-
-        # flush all the canvas operation
-        Builder.sync()
-
-        # tick before draw
-        Clock.tick_draw()
-
-        # flush all the canvas operation
-        Builder.sync()
-
-        window = self.window
-        if window and window.canvas.needs_redraw:
-            window.dispatch('on_draw')
-            window.dispatch('on_flip')
-
-        # don't loop if we don't have listeners !
-        if len(self.event_listeners) == 0:
-            Logger.error('Base: No event listeners have been created')
-            Logger.error('Base: Application will leave')
-            self.exit()
-            return False
-
-        return self.quit
-
     def run(self):
         '''Main loop'''
         while not self.quit:
             self.idle()
         self.exit()
-
-    def async_run(self):
-        loop = asyncio.get_event_loop()
-        loop.run_until_complete(self.async_mainloop())
-        loop.close()
 
     def exit(self):
         '''Close the main loop and close the window.'''
@@ -565,25 +514,6 @@ def runTouchApp(widget=None, slave=False):
             EventLoop.mainloop()
         elif opt == 'async':
             EventLoop.async_run()
-    finally:
-        stopTouchApp()
-
-
-@async_coroutine
-def async_runTouchApp(widget=None, slave=False):
-    '''Identical to :func:`runTouchApp` but instead is a coroutine
-    that can be run in an existing async event loop.
-
-    .. versionadded:: 1.10.1
-    '''
-    _runTouchApp_prepare(widget=widget, slave=slave)
-
-    # we are in a slave mode, don't do dispatching.
-    if slave:
-        return
-
-    try:
-        yield from EventLoop.async_mainloop()
     finally:
         stopTouchApp()
