+++ kivy/clock.py
@@ -382,6 +382,24 @@ try:
 except ImportError:
     asyncio = None
 
+if asyncio:
+    from kivy.clock_async import AsyncClockBaseBehavior, \
+        AsyncClockBaseFreeInterruptOnly, AsyncClockBaseInterruptBehavior
+else:
+    from kivy.compat import PY3CompatCls
+
+
+    class AsyncClockBaseBehavior(PY3CompatCls):
+        pass
+
+
+    class AsyncClockBaseFreeInterruptOnly(PY3CompatCls):
+        pass
+
+
+    class AsyncClockBaseInterruptBehavior(PY3CompatCls):
+        pass
+
 # some reading: http://gameprogrammingpatterns.com/game-loop.html
 
 
@@ -470,7 +488,7 @@ except (OSError, ImportError, AttributeError):
         time.sleep(microseconds / 1000000.)
 
 
-class ClockBaseBehavior(object):
+class ClockBaseBehavior(AsyncClockBaseBehavior):
     '''The base of the kivy clock.
     '''
 
@@ -603,41 +621,6 @@ class ClockBaseBehavior(object):
 
         return self._dt
 
-    @async_coroutine
-    def async_idle(self):
-        '''(internal) async version of :meth:`idle`.
-        '''
-        fps = self._max_fps
-        if fps > 0:
-            min_sleep = self.get_resolution()
-            undershoot = 4 / 5. * min_sleep
-            ready = self._check_ready
-
-            slept = False
-            done, sleeptime = ready(fps, min_sleep, undershoot)
-            while not done:
-                slept = True
-                yield from asyncio.sleep(sleeptime)
-                done, sleeptime = ready(fps, min_sleep, undershoot)
-
-            if not slept:
-                yield from asyncio.sleep(0)
-        else:
-            yield from asyncio.sleep(0)
-
-        current = self.time()
-        self._dt = current - self._last_tick
-        self._last_tick = current
-        return current
-
-    @async_coroutine
-    def async_tick(self):
-        '''async version of :meth:`tick`. '''
-        self.pre_idle()
-        ts = self.time()
-        current = yield from self.async_idle()
-        self.post_idle(ts, current)
-
     def tick_draw(self):
         '''Tick the drawing counter.
         '''
@@ -674,7 +657,8 @@ ClockBaseBehavior.time.__doc__ = \
     '''Proxy method for :func:`~kivy.compat.clock`. '''
 
 
-class ClockBaseInterruptBehavior(ClockBaseBehavior):
+class ClockBaseInterruptBehavior(
+        AsyncClockBaseInterruptBehavior, ClockBaseBehavior):
     '''A kivy clock which can be interrupted during a frame to execute events.
     '''
 
@@ -694,12 +678,6 @@ class ClockBaseInterruptBehavior(ClockBaseBehavior):
     def usleep(self, microseconds):
         self._event.clear()
         self._event.wait(microseconds / 1000000.)
-
-    @async_coroutine
-    def async_usleep(self, microseconds):
-        self._async_event.clear()
-        yield from asyncio.wait_for(
-            self._async_event.wait(), microseconds / 1000000.)
 
     def on_schedule(self, event):
         fps = self._max_fps
@@ -724,32 +702,6 @@ class ClockBaseInterruptBehavior(ClockBaseBehavior):
                 fps, resolution, 4 / 5. * resolution, event)
             if not done:
                 event.wait(sleeptime)
-
-        current = self.time()
-        self._dt = current - self._last_tick
-        self._last_tick = current
-        event.clear()
-        # anything scheduled from now on, if scheduled for the upcoming frame
-        # will cause a timeout of the event on the next idle due to on_schedule
-        # `self._last_tick = current` must happen before clear, otherwise the
-        # on_schedule computation is wrong when exec between the clear and
-        # the `self._last_tick = current` bytecode.
-        return current
-
-    @async_coroutine
-    def async_idle(self):
-        fps = self._max_fps
-        event = self._async_event
-        resolution = self.get_resolution()
-        if fps > 0:
-            done, sleeptime = self._check_ready(
-                fps, resolution, 4 / 5. * resolution, event)
-            if not done:
-                yield from asyncio.wait_for(event.wait(), sleeptime)
-            else:
-                yield from asyncio.sleep(0)
-        else:
-            yield from asyncio.sleep(0)
 
         current = self.time()
         self._dt = current - self._last_tick
@@ -826,7 +778,8 @@ class ClockBaseFreeInterruptAll(
 
 
 class ClockBaseFreeInterruptOnly(
-        ClockBaseInterruptFreeBehavior, CyClockBaseFree):
+        AsyncClockBaseFreeInterruptOnly, ClockBaseInterruptFreeBehavior,
+        CyClockBaseFree):
     '''The ``free_only`` kivy clock. See module for details.
     '''
 
@@ -862,53 +815,6 @@ class ClockBaseFreeInterruptOnly(
                     event.wait(sleeptime - undershoot)
                 current = self.time()
                 sleeptime = 1 / fps - (current - self._last_tick)
-
-        self._dt = current - self._last_tick
-        self._last_tick = current
-        event.clear()  # this needs to stay after _last_tick
-        return current
-
-    @async_coroutine
-    def async_idle(self):
-        fps = self._max_fps
-        current = self.time()
-        event = self._async_event
-        if fps > 0:
-            min_sleep = self.get_resolution()
-            usleep = self.usleep
-            undershoot = 4 / 5. * min_sleep
-            min_t = self.get_min_free_timeout
-            interupt_next_only = self.interupt_next_only
-
-            sleeptime = 1 / fps - (current - self._last_tick)
-            slept = False
-            while sleeptime - undershoot > min_sleep:
-                if event.is_set():
-                    do_free = True
-                else:
-                    t = min_t()
-                    if not t:
-                        do_free = True
-                    elif interupt_next_only:
-                        do_free = False
-                    else:
-                        sleeptime = min(sleeptime, t - current)
-                        do_free = sleeptime - undershoot <= min_sleep
-
-                if do_free:
-                    event.clear()
-                    self._process_free_events(current)
-                else:
-                    slept = True
-                    yield from asyncio.wait_for(
-                        event.wait(), sleeptime - undershoot)
-                current = self.time()
-                sleeptime = 1 / fps - (current - self._last_tick)
-
-            if not slept:
-                yield from asyncio.sleep(0)
-        else:
-            yield from asyncio.sleep(0)
 
         self._dt = current - self._last_tick
         self._last_tick = current
